dia6/deploy_render/requirements.txt
django==3.2
djangorestframework
mysqlclient
django-cors-headers
cloudinary
djangorestframework-simplejwt
python-decouple
gunicorn
whitenoise
drf-yasg
